Move _comp_df encoding logic out of base into concrete subclasses
